factor out async tg
